feat: can edit and delete comments from threads #298
