+++ frontend/lib/components/appbar.dart
@@ -2,12 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:frontend/components/SearchResultWidget.dart';
 import 'package:frontend/components/deckTheme.dart';
 import 'package:frontend/components/eventNotifier.dart';
-import 'package:frontend/components/SearchResultWidget.dart';
-import 'package:frontend/main.dart';
 import 'package:frontend/models/event.dart';
-import 'package:frontend/routes/company/CompanyScreen.dart';
-import 'package:frontend/routes/member/MemberScreen.dart';
-import 'package:frontend/routes/speaker/SpeakerScreen.dart';
 import 'package:frontend/services/eventService.dart';
 import 'package:provider/provider.dart';
 import 'package:frontend/models/company.dart';
