+++ frontend/lib/routes/company/CompanyScreen.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/components/EditableCard.dart';
-import 'package:frontend/components/addThreadForm.dart';
+import 'package:frontend/components/threads/addThreadForm.dart';
 import 'package:frontend/components/appbar.dart';
 import 'package:frontend/components/deckTheme.dart';
 import 'package:frontend/components/eventNotifier.dart';
 import 'package:frontend/components/participationCard.dart';
-import 'package:frontend/components/threadCard.dart';
+import 'package:frontend/components/threads/participations/communicationsList.dart';
+import 'package:frontend/components/threads/threadCard.dart';
 import 'package:frontend/models/company.dart';
 import 'package:frontend/routes/company/CompanyTableNotifier.dart';
 import 'package:frontend/routes/company/EditCompanyForm.dart';
@@ -181,6 +182,8 @@ class _CompanyScreenState extends State<CompanyScreen>
                       ),
                       CommunicationsList(
                           participations: widget.company.participations ?? [],
+                          id: widget.company.id,
+                          type: CommunicationType.COMPANY,
                           small: small),
                     ]),
                   ),
