+++ frontend/lib/routes/meeting/AddMeetingMemberForm.dart
@@ -86,7 +86,6 @@ class _AddMeetingMemberForm extends State<AddMeetingMemberForm> {
                 builder: (context, snapshot) {
                   if (snapshot.hasData) {
                     List<Member> membsMatched = snapshot.data as List<Member>;
-                    print("membs matched:" + membsMatched.toString());
                     return searchResults(membsMatched, height);
                   } else {
                     return Center(child: CircularProgressIndicator());
