+++ frontend/lib/routes/speaker/SpeakerScreen.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/components/EditableCard.dart';
-import 'package:frontend/components/addThreadForm.dart';
+import 'package:frontend/components/threads/addThreadForm.dart';
 import 'package:frontend/components/appbar.dart';
 import 'package:frontend/components/deckTheme.dart';
 import 'package:frontend/components/eventNotifier.dart';
 import 'package:frontend/components/participationCard.dart';
 import 'package:frontend/components/router.dart';
-import 'package:frontend/components/threadCard.dart';
+import 'package:frontend/components/threads/participations/communicationsList.dart';
+import 'package:frontend/components/threads/threadCard.dart';
 import 'package:frontend/routes/speaker/speakerNotifier.dart';
 import 'package:frontend/components/status.dart';
 import 'package:frontend/main.dart';
@@ -125,6 +126,8 @@ class _SpeakerScreenState extends State<SpeakerScreen>
                   ),
                   CommunicationsList(
                       participations: widget.speaker.participations ?? [],
+                      id: widget.speaker.id,
+                      type: CommunicationType.SPEAKER,
                       small: small),
                 ]),
               ),
